+++ frontend/src/components/PairRoom/TodoListCard/TodoItem/TodoItem.tsx
@@ -3,19 +3,19 @@ import CheckBox from '@/components/common/CheckBox/CheckBox';
 import * as S from './TodoItem.styles';
 
 interface TodoItemProps {
-  position: number;
+  id: number;
   content: string;
   onDragStart: (position: number) => void;
   onDragEnter: (position: number) => void;
   onDrop: (event: React.DragEvent<HTMLDivElement>) => void;
 }
 
-const TodoItem = ({ position, content, onDragStart, onDragEnter, onDrop }: TodoItemProps) => {
+const TodoItem = ({ id, content, onDragStart, onDragEnter, onDrop }: TodoItemProps) => {
   return (
     <S.Layout
       draggable
-      onDragStart={() => onDragStart(position)}
-      onDragEnter={() => onDragEnter(position)}
+      onDragStart={() => onDragStart(id)}
+      onDragEnter={() => onDragEnter(id)}
       onDragOver={(event) => event.preventDefault()}
       onDrop={onDrop}
     >
