+++ frontend/src/components/PairRoom/TodoListCard/TodoListCard.tsx
@@ -1,4 +1,4 @@
-import { useState, useRef } from 'react';
+import { useState } from 'react';
 
 import { IoIosCheckbox } from 'react-icons/io';
 import { LuPlus } from 'react-icons/lu';
@@ -8,6 +8,7 @@ import Input from '@/components/common/Input/Input';
 import { PairRoomCard } from '@/components/PairRoom/PairRoomCard';
 import TodoItem from '@/components/PairRoom/TodoListCard/TodoItem/TodoItem';
 
+import useDragAndDrop from '@/hooks/common/useDragAndDrop';
 import useInput from '@/hooks/common/useInput';
 
 import { theme } from '@/styles/theme';
@@ -15,39 +16,16 @@ import { theme } from '@/styles/theme';
 import * as S from './TodoListCard.styles';
 
 const TodoListCard = () => {
-  const [isInputOpen, setIsInputOpen] = useState(false);
   const [todos, setTodos] = useState(['Item 1', 'Item 2', 'Item 3', 'Item 4', 'Item 5']);
+  const [isInputOpen, setIsInputOpen] = useState(false);
+
+  const handleTodos = (newTodos: string[]) => setTodos(newTodos);
 
   const { value, handleChange, resetValue } = useInput();
-
-  const dragItem = useRef<number | null>(null);
-  const dragOverItem = useRef<number | null>(null);
-
-  const handleDragStart = (position: number) => (dragItem.current = position);
-
-  const handleDragEnter = (position: number) => (dragOverItem.current = position);
-
-  const handleDrop = (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault();
-
-    if (dragItem.current === null || dragOverItem.current === null) return;
-
-    const newTodos = [...todos];
-    console.log(dragItem.current, newTodos[dragItem.current]);
-    const draggedItem = newTodos[dragItem.current];
-
-    newTodos.splice(dragItem.current, 1);
-    newTodos.splice(dragOverItem.current, 0, draggedItem);
-
-    dragItem.current = null;
-    dragOverItem.current = null;
-
-    setTodos(newTodos);
-  };
+  const { handleDragStart, handleDragEnter, handleDrop } = useDragAndDrop(todos, handleTodos);
 
   const addTodoItem = () => {
     setTodos((prev) => [...prev, value]);
-
     resetValue();
     setIsInputOpen(false);
   };
@@ -59,7 +37,7 @@ const TodoListCard = () => {
         {todos.map((todo, idx) => (
           <TodoItem
             key={idx}
-            position={idx}
+            id={idx}
             content={todo}
             onDragStart={handleDragStart}
             onDragEnter={handleDragEnter}
